.kshrc
Commented out the completion for man cause it was noticeably slowing down
shell startup.
Don't need the '-v' option, also not portable.
Said HISTSIZE default is 500 but it seemed to default to 1000?
Also I didn't see it in the manpage but it removes at 125% of HISTSIZE not
at the HISTSIZE.
https://marc.info/?l=openbsd-tech&m=131490865525379&w=2
https://github.com/openbsd/src/blob/master/bin/ksh/history.c#L842